fix: Fix BFloat16 to NumPy conversion error in Stage 1 adapter training
- Added .float() conversion before .numpy() in EmbeddingCompressor.fit()
- Resolves TypeError when using PCA with BFloat16 model embeddings
- Updated LOG.md with issue analysis and fix details
- Ready for Stage 1 re-run on H100s

+++ train_adapter_only.py
@@ -36,7 +36,8 @@ class EmbeddingCompressor:
         """Fit compression on sample embeddings"""
         if self.method == "pca":
             # Use PCA for compression
-            embeddings_flat = embeddings.reshape(-1, self.input_dim).cpu().numpy()
+            # Convert BFloat16 to Float32 for numpy compatibility
+            embeddings_flat = embeddings.reshape(-1, self.input_dim).cpu().float().numpy()
             pca = PCA(n_components=self.output_dim)
             pca.fit(embeddings_flat[:10000])  # Fit on subset
